requirements.txt
Django>=5.2
djangorestframework>=3.16
black>=25.1
flake8>=7.2
isort>=6.0